telegram_ai_bot/requirements.txt
python-telegram-bot==20.3
torch
transformers
ftfy
tqdm
scipy
numpy
Pillow